user token update and delete protected routes and crud actions

+++ backend/controllers/itemController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Item = require("../models/itemModel");
+const User = require("../models/userModel");
 
 // @desc: Get all items
 // @route: GET /api/items
 // @access: Public
 const getItems = asyncHandler(async (req, res) => {
-  const items = await Item.find({});
+  const items = await Item.find({ user: req.user.id });
 
   res.status(200).json(items);
 });
@@ -41,6 +42,7 @@ const createItem = asyncHandler(async (req, res) => {
   }
 
   const item = await Item.create({
+    user: req.user.id,
     name: req.body.name,
     price: req.body.price,
     description: req.body.description,
@@ -63,6 +65,19 @@ const updateItem = asyncHandler(async (req, res) => {
     throw new Error("Item not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  if(!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  // Check if user is authorized to update item
+  if(item.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedItem = await Item.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -80,6 +95,19 @@ const deleteItem = asyncHandler(async (req, res) => {
   if (!item) {
     res.status(404);
     throw new Error("Item not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if(!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  // Check if user is authorized to update item
+  if(item.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await item.deleteOne();
